LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-named-scroll-progress-timeline.tentative-expected.txt


FAIL scroll-timeline-name is referenceable in animation-timeline on the declaring element itself assert_equals: expected "100px" but got "50px"
FAIL scroll-timeline-name is referenceable in animation-timeline on that element's descendants assert_equals: expected "100px" but got "50px"
PASS scroll-timeline-name is not referenceable in animation-timeline on that element's siblings
FAIL scroll-timeline-name on an element which is not a scroll-container assert_equals: Animation with an unresolved current time expected "none" but got "50px"
FAIL Change in scroll-timeline-name to match animation timeline updates animation. assert_equals: expected null but got object "[object DocumentTimeline]"
FAIL Change in scroll-timeline-name to no longer match animation timeline updates animation. assert_equals: expected "75px" but got "50px"
FAIL Timeline lookup updates candidate when closer match available. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'anim.timeline.source.id')"
FAIL Timeline lookup updates candidate when match becomes available. assert_equals: expected "75px" but got "50.330002px"
FAIL scroll-timeline-axis is block assert_equals: expected "50" but got "0"
FAIL scroll-timeline-axis is inline assert_equals: expected "50" but got "0"
FAIL scroll-timeline-axis is x assert_equals: expected "50" but got "0"
FAIL scroll-timeline-axis is y assert_equals: expected "50" but got "0"
FAIL scroll-timeline-axis is mutated assert_equals: expected "25" but got "0"
